pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "mortality"
version = "0.1.0"
description = "Multi-agent mortality experiments with time-aware LLM orchestration"
readme = "README.md"
authors = [{name = "Mortality Lab", email = "lab@example.com"}]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
  "anyio>=4.0,<5.0",
  "httpx>=0.27.0,<0.28.0",
  "pydantic>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",
  "typing-extensions>=4.9,<5.0"
]

[project.optional-dependencies]
openai = ["openai>=1.42.0"]
anthropic = ["anthropic>=0.34.0"]
grok = ["xai-sdk>=0.2.0"]
gemini = ["google-genai>=0.3.0"]
test = ["pytest>=8.3"]

[tool.hatch.metadata]
allow-direct-references = true
